V 021.09.05 easy to use uniform buffer easy game implementation transformation stack

+++ Project ORKA/Uniforms.hpp
@@ -5,20 +5,21 @@
 
 struct GlobalUniformData {
 	//values have to be layed out in blocks of 16 bytes in memory
-	Matrix mMatrix = Matrix(1);			//16 * 4
-	Matrix vpMatrix = Matrix(1);		//16 * 4
-	Vec4 worldOffset = Vec4(0,0,0,0);	//16
+	//data has to match up with "uniforms.glsl"
+	Matrix mMatrix = Matrix(1);						//16 * 4
+	Matrix vpMatrix = Matrix(1);					//16 * 4
+	Vec4 worldOffset = Vec4(0, 0, 0, 0);			//16
 	Vec4 cameraVector = Vec4(0, 0, 0, 0);			//16
 	Vec4 chunkOffsetVector = Vec4(0, 0, 0, 0);		//16
 	Vec4 customColor = Vec4(0, 0, 0, 1);
-	float time = 0.0f;					//16
+	float time = 0.0f;								//16
 	float custom1 = 0.0f;
 	float custom2 = 0.0f;
 	float custom3 = 0.0f;
-	Int distortion = 0;				//16
-	float placeholder1 = 0.0f;
-	float placeholder2 = 0.0f;
-	float placeholder3 = 0.0f;
+	Int distortion = 0;								//16
+	Int placeholder1 = 0.0f;
+	Int placeholder2 = 0.0f;
+	Int placeholder3 = 0.0f;
 };
 
 struct Uniforms {
